+++ src/Adafruit_PWMServoDriver.h
@@ -48,7 +48,7 @@ class Adafruit_PWMServoDriver {
   Adafruit_PWMServoDriver(uint8_t addr = 0x40);
   void begin(void);
   void reset(void);
-  void setPWMFreq(float freq);
+  void setPWMFreq(int freq);
   void setPWM(uint8_t num, uint16_t on, uint16_t off);
   void setPin(uint8_t num, uint16_t val, bool invert=false);
   int getPWM(uint8_t num);
